Filter Callsigns with more than 6 characters and locators with more than 4.

+++ Inc/Display.h
@@ -39,7 +39,7 @@ void show_variable(uint16_t x, uint16_t y, int variable);
 
 void show_short(uint16_t x, uint16_t y, uint8_t variable);
 
-void show_wide(uint16_t x, uint16_t y, int variable);
+void show_wide(uint16_t x, uint16_t y, uint16_t variable);
 
 void show_UTC_time(uint16_t x, uint16_t y, int utc_hours, int utc_minutes,
 		int utc_seconds, int color);
